rename redoes_utils.py to redeos_utils.py
