+++ tests/redeos_utils.py
